requirements.txt
djangorestframework-jwt==1.11.0